fix: eval task cost_usd serialized as string instead of number
EvalTaskResponse.cost_usd was Decimal | None — Pydantic v2 serializes
Decimal as string ("0.0012") not number (0.0012) in JSON. Frontend
Number(t.cost_usd).toFixed(4) happened to work because Number()
parses strings, but this is fragile and inconsistent with
EvalUsageResponse which uses float.

Changed to float | None to match the rest of the API.

+++ apps/api/app/schemas/evaluation.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 
 from pydantic import BaseModel, field_validator
 
@@ -32,7 +31,7 @@ class EvalTaskResponse(BaseModel):
     llm_model: str | None
     input_tokens: int | None
     output_tokens: int | None
-    cost_usd: Decimal | None
+    cost_usd: float | None
     duration_ms: int | None
     retries: int
     started_at: datetime | None
